fix: uds websockets and independent kill switch

+++ client/src/commonMain/kotlin/com/zaneschepke/wireguardautotunnel/client/data/entity/TunnelConfig.kt
@@ -5,12 +5,12 @@ import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverters
-import com.zaneschepke.wireguardautotunnel.client.data.AppKeyringConverter
+import com.zaneschepke.wireguardautotunnel.client.data.converter.AppKeyringConverter
 import com.zaneschepke.wireguardautotunnel.client.data.model.EncryptedField
 
 @Entity(tableName = "tunnel_config", indices = [Index(value = ["name"], unique = true)])
 data class TunnelConfig(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     @ColumnInfo(name = "name") val name: String,
     @field:TypeConverters(AppKeyringConverter::class)
     @ColumnInfo(name = "quick_config") val quickConfig: EncryptedField,
